Enhanced swagger, refactored services, deleted holiday service
