+++ stock-service/src/main/java/com/banka1/stock_service/controller/StockMarketDataAdminController.java
@@ -2,6 +2,7 @@ package com.banka1.stock_service.controller;
 
 import com.banka1.stock_service.dto.StockMarketDataRefreshResponse;
 import com.banka1.stock_service.service.StockMarketDataRefreshService;
+import io.swagger.v3.oas.annotations.Operation;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -25,6 +26,7 @@ public class StockMarketDataAdminController {
      * @param ticker stock ticker to refresh
      * @return summary of the completed refresh operation
      */
+    @Operation(summary = "Refresh stock market data by ticker")
     @PostMapping("/{ticker}/refresh-market-data")
     @PreAuthorize("hasAnyRole('ADMIN', 'SUPERVISOR')")
     public StockMarketDataRefreshResponse refreshStockMarketData(@PathVariable String ticker) {
