+++ stock-service/src/main/java/com/banka1/stock_service/runner/ForexPairApiImportService.java
@@ -1,4 +1,4 @@
-package com.banka1.stock_service.service;
+package com.banka1.stock_service.runner;
 
 import com.banka1.stock_service.client.AlphaVantageClient;
 import com.banka1.stock_service.domain.ForexPair;
