+++ stock-service/src/main/java/com/banka1/stock_service/runner/StockTickerSeedService.java
@@ -1,4 +1,4 @@
-package com.banka1.stock_service.service;
+package com.banka1.stock_service.runner;
 
 import com.banka1.stock_service.domain.Listing;
 import com.banka1.stock_service.domain.ListingType;
