+++ stock-service/src/main/java/com/banka1/stock_service/scheduler/ListingMarketDataScheduler.java
@@ -1,4 +1,4 @@
-package com.banka1.stock_service.service;
+package com.banka1.stock_service.scheduler;
 
 import com.banka1.stock_service.config.ListingRefreshProperties;
 import com.banka1.stock_service.domain.Listing;
@@ -6,6 +6,8 @@ import com.banka1.stock_service.domain.ListingType;
 import com.banka1.stock_service.dto.ListingRefreshBatchResponse;
 import com.banka1.stock_service.dto.StockExchangeStatusResponse;
 import com.banka1.stock_service.repository.ListingRepository;
+import com.banka1.stock_service.service.ListingMarketDataRefreshService;
+import com.banka1.stock_service.service.StockExchangeService;
 import jakarta.annotation.PostConstruct;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
